More review comments + formatting
