--- keras_core/utils/generic_utils.py
@@ -1,15 +1,0 @@
-def to_list(x):
-    """Normalizes a list/tensor into a list.
-
-    If a tensor is passed, we return
-    a list of size 1 containing the tensor.
-
-    Args:
-        x: target object to be normalized.
-
-    Returns:
-        A list.
-    """
-    if isinstance(x, list):
-        return x
-    return [x]
